backend/requirements.txt
# FastAPI Requirements (requirements.txt)
fastapi
uvicorn
python-multipart
pydantic
paramiko
hashlib
pefile
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
psycopg2-binary
python-dotenv
aiofiles
prometheus-fastapi-instrumentor
